requirements.txt
flask
pyodbc
python-dotenv
flask-cors